requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.1
langgraph>=0.0.20
langsmith>=0.0.72
python-dotenv>=1.0.0
pydantic>=2.5.0
langserve>=0.0.30
openai>=1.12.0
requests>=2.31.0
chromadb>=0.4.18 